+++ lib/punch/plugins.rb
@@ -1,13 +1,16 @@
 # frozen_string_literal: true
 
+require "logger"
 require_relative "plugins/playbox"
 require_relative "plugins/preview"
-require_relative "plugins/logger"
 
 module Punch
 
   PlayboxPlug = Playbox.plugin
   PreviewPlug = Preview.plugin
+
+  Logger.extend(Plugin)
   LoggerPlug  = Logger.plugin
+  LoggerPlug.object = Logger.new(IO::NULL)
 
 end